lots of changes
I added the wireframes of the new pages. I made some additions to the database. I almost finished (I hope :d) the file uploads outside of the applications that is made by applying to an announcement. I also made additions to the internships that starts by applying to an announcement.